Rename index.js to direct.js and export.
